04/03/2019
Unraveled a CF and made the code to implement a new SQLite database everytime someone downloads the app. I also made functionality that populates the database with the proper tables and columns. I made functionality to add an entry, find an entry, delete an entry, update an entry, and load the handlers in all of the tables.